[M148] [Actor Login M2] Add a toast for failed remove permission attempts
Original change's description:
> [Actor Login M2] Add a toast for failed remove permission attempts
>
> Recording: http://shortn/_kK5aIWH4fm
>
> This adds a callback to ActorLoginPermissionsManager's revoke method
> that will contain the status of the delete request. Currently only
> federated permission deletion can fail. The status is propagated all the
> way to the UI code and there it's used to show an error toast.
>
> Revoking a non-existent federated permission is also considered an
> error, so I'm adding a check to only send delete requests to the
> permissions service when we know the permission exists.
>
> Fixed: b:497710964
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7727221
> Commit-Queue: Oleksandr Tara <otara@google.com>
> Cr-Commit-Position: refs/heads/main@{#1613823}

(cherry picked from commit 97943e487f52a3ac8b28618c2efa7b8a740c3f68)

Bug: 502456540,b:497710964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7760140
Commit-Queue: Oleksandr Tara <otara@google.com>
Cr-Commit-Position: refs/branch-heads/7778@{#801}
Cr-Branched-From: 77f495ee216d4c3cc784d33658bad4778c0680ee-refs/heads/main@{#1610480}

+++ components/password_manager/core/browser/actor_login/actor_login_permissions_manager_impl.cc
@@ -4,6 +4,8 @@
 
 #include "components/password_manager/core/browser/actor_login/actor_login_permissions_manager_impl.h"
 
+#include <algorithm>
+
 #include "base/functional/bind.h"
 #include "base/functional/callback.h"
 #include "base/strings/utf_string_conversions.h"
@@ -94,16 +96,32 @@ void ActorLoginPermissionsManagerImpl::RemoveObserver(
 
 void ActorLoginPermissionsManagerImpl::RevokePermission(
     const std::string& signon_realm,
-    const std::string& username) {
+    const std::string& username,
+    base::OnceCallback<void(bool)> callback) {
   presenter_.RevokeActorLoginPermission(signon_realm, username);
   // The service is constructed via a factory that returns a nullptr for
   // incognito and guest profiles. The settings page is not accessible for those
   // profiles, so we can assert that the service is valid.
   CHECK(actor_login_permission_service_);
-  actor_login_permission_service_->DeletePermission(
-      url::Origin::Create(GURL(signon_realm)), username,
-      base::BindOnce(&ActorLoginPermissionsManagerImpl::OnPermissionDeleted,
-                     weak_ptr_factory_.GetWeakPtr()));
+
+  if (std::ranges::any_of(
+          last_federated_permissions_,
+          [&](const FederatedPermission& permission) {
+            return permission.rp_embedder_origin.GetURL().spec() ==
+                       signon_realm &&
+                   permission.chosen_account_email == username;
+          })) {
+    actor_login_permission_service_->DeletePermission(
+        url::Origin::Create(GURL(signon_realm)), username,
+        base::BindOnce(&ActorLoginPermissionsManagerImpl::OnPermissionDeleted,
+                       weak_ptr_factory_.GetWeakPtr(), std::move(callback)));
+  } else {
+    // If there is no corresponding federated permission, skip the request.
+    // Delete requests for non-existent permissions result in an error, which
+    // would trigger the error toast in the UI.
+    base::SequencedTaskRunner::GetCurrentDefault()->PostTask(
+        FROM_HERE, base::BindOnce(std::move(callback), true));
+  }
 }
 
 void ActorLoginPermissionsManagerImpl::GetAllPermissions(
@@ -136,6 +154,7 @@ void ActorLoginPermissionsManagerImpl::OnFederatedPermissionsReceived(
   base::flat_set<password_manager::ActorLoginPermission>
       saved_passwords_permissions =
           presenter_.GetActorLoginPermissions(sync_service);
+  last_federated_permissions_ = federated_permissions;
   std::move(callback).Run(
       MergePermissions(std::move(saved_passwords_permissions),
                        std::move(federated_permissions)));
@@ -147,10 +166,13 @@ void ActorLoginPermissionsManagerImpl::NotifyObservers() {
   }
 }
 
-void ActorLoginPermissionsManagerImpl::OnPermissionDeleted(bool success) {
+void ActorLoginPermissionsManagerImpl::OnPermissionDeleted(
+    base::OnceCallback<void(bool)> callback,
+    bool success) {
   if (success) {
     NotifyObservers();
   }
+  std::move(callback).Run(success);
 }
 
 void ActorLoginPermissionsManagerImpl::OnSavedPasswordsPresenterInitialized() {
